+++ data.py
@@ -11,7 +11,7 @@ def returnLanguage(lg, stat):
 def createGuild(guild_id, role_bot):
     with open("database/{}.json".format(guild_id), "x") as outfile:
         json.dump(
-            {"botID": role_bot, "prefix": ".Check ", "mp": True, "language": "en", "sysMessages": True, "admin": []},
+            {"botID": role_bot, "prefix": ".Check ", "mp": True, "language": "en", "sysMessages": True, "admin": [],"teacher":[]},
             outfile)
 
 
@@ -32,7 +32,7 @@ def readGuild(guild_id):
 def get_prefix(client, message):
     with open("database/{}.json".format(message.guild.id), "r") as outfile:
         var = json.load(outfile)
-    return var["prefix"]
+    return ['.Check ', var["prefix"]]
 
 
 def set_prefix(guild_id, prefix):